feat: extract stateless utilities to DIRACCommon package
Extract ReturnValues and DErrno to separate DIRACCommon package to solve
circular dependencies where DiracX needs DIRAC utilities without global
state initialization.

- Create DIRACCommon package with stateless utilities
- Maintain backward compatibility in original DIRAC locations
- Add independent testing and CI workflows
- Configure automated deployment for both packages
