+++ .github/workflows/deployment.yml
@@ -106,8 +106,19 @@ jobs:
               fi
             fi
           fi
-      - name: Build distributions
+      - name: Build DIRACCommon distribution
+        if: steps.check-tag.outputs.create-release == 'true'
         run: |
+          cd dirac-common
+          python -m build
+          cd ..
+      - name: Pin DIRACCommon version and build DIRAC distribution
+        run: |
+          # If we're making a release, pin DIRACCommon to exact version
+          if [[ "${{ steps.check-tag.outputs.create-release }}" == "true" ]]; then
+            DIRACCOMMON_VERSION=$(cd dirac-common && python -m setuptools_scm | sed 's@Guessed Version @@g' | sed -E 's@(\.dev|\+g).+@@g')
+            python .github/workflows/pin_diraccommon_version.py "$DIRACCOMMON_VERSION"
+          fi
           python -m build
       - name: Make release on GitHub
         if: steps.check-tag.outputs.create-release == 'true'
@@ -123,7 +134,14 @@ jobs:
             --version="${NEW_VERSION}" \
             --rev="$(git rev-parse HEAD)" \
             --release-notes-fn="release.notes.new"
-      - name: Publish package on PyPI
+      - name: Publish DIRACCommon to PyPI
+        if: steps.check-tag.outputs.create-release == 'true'
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          user: __token__
+          password: ${{ secrets.PYPI_API_TOKEN }}
+          packages-dir: dirac-common/dist/
+      - name: Publish DIRAC to PyPI
         if: steps.check-tag.outputs.create-release == 'true'
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
