+++ dirac-common/src/DIRACCommon/Utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+DIRACCommon.Utils - Stateless utility functions
+"""
