+++ dirac-common/tests/__init__.py
@@ -1,0 +1,1 @@
+"""DIRACCommon test suite"""
