+++ setup.cfg
@@ -30,6 +30,7 @@ install_requires =
     certifi
     cwltool
     diraccfg
+    DIRACCommon
     diracx-client >=v0.0.1a18
     diracx-core >=v0.0.1a18
     db12
